+++ CHANGELOG.md
@@ -1,4 +1,20 @@
 # 更新日志\n
+## [ForwardWidgets-huangxd] - 2025-09-03 02:45:54
+- 仓库: https://github.com/huangxd-/ForwardWidgets
+- 分支: main
+- 新增: 47cb3ce52cb083a1faa1b33a841ac543bc118a7e
+
+## [ForwardWidgets-2kuai] - 2025-09-03 02:45:54
+- 仓库: https://github.com/2kuai/ForwardWidgets
+- 分支: main
+- 新增: dfda2b353555c52478c92776f8290bf69a44eaca
+
+## [ForwardWidgets-pack1r] - 2025-09-03 02:45:54
+- 仓库: https://github.com/pack1r/ForwardWidgets
+- 分支: main
+- 新增: 42f40bad2a1953900f45253d66da75a37dceb6ee
+
+
 ## [ForwardWidgets-huangxd] - 2025-09-02 02:59:44
 - 仓库: https://github.com/huangxd-/ForwardWidgets
 - 分支: main